TechChallenge4/requirements.txt
# Core
opencv-python>=4.8.0
numpy>=1.24.0

# YOLO e Deep Learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# GPU Support (opcional, instalar se tiver NVIDIA GPU)
# Para CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# Para CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Processamento de Imagem
pillow>=10.0.0

# UI/Progress
tqdm>=4.65.0
rich>=13.0.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilitários
pyyaml>=6.0
python-dateutil>=2.8.0

# Detecção de Faces e Emoções
deepface>=0.0.79
tf-keras>=2.16.0
tensorflow>=2.16.0
